% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{medical_data}
\alias{medical_data}
\title{Load Infertility Dataset}
\usage{
medical_data()
}
\value{
A data.frame containing infertility cases with labeled predictors suitable for logistic regression
}
\description{
Load Infertility Dataset
}
